Masked SlackFormatService behind an interface
